refactor(models): Split multi-class files into ONEX single-class-per-file standard
- Split model_resource_manager.py into model_circuit_breaker_config.py and
  model_circuit_breaker_stats_response.py
- Split model_pii.py into 5 files (model_pii_type.py, model_pii_match.py,
  model_pii_detection_result.py, model_pii_detector_config.py,
  model_pii_pattern_config.py)
- Split model_retry.py into 3 files (model_retry_config.py,
  model_retry_attempt_info.py, model_retry_statistics.py)
- Split model_observability.py into model_structured_log_entry.py and
  model_correlation_context.py
- Split model_health.py into 6 files (model_health_status.py,
  model_health_check_config.py, model_health_check_details.py,
  model_health_check_result.py, model_resource_health_check.py,
  model_system_health.py)
- Split model_graph_memory.py into 6 files (model_memory_connection.py,
  model_related_memory.py, model_related_memory_result.py,
  model_connections_result.py, model_graph_memory_health.py,
  model_graph_memory_config.py)
- Fix TC001 linting errors by moving type-only imports into TYPE_CHECKING blocks
- Remove all backward compatibility aliases per project policy
- Fix naming validator to allow model_*.py in models/adapters/ directories
